store_consumer/requirements.txt
pandas
numpy
scikit-learn
scipy
sqlite3
kafka-python
kafka-python-ng
websocket
python-socketio
websocket-client>=1.6.0
requests
python-dotenv